Consider only the last transaction
