+++ src/Provider/Payone/EventListener/InvoiceRendererEventListener.php
@@ -6,10 +6,10 @@ namespace PayonePayment\Provider\Payone\EventListener;
 
 use PayonePayment\DataAbstractionLayer\Aggregate\PayonePaymentOrderTransactionDataEntity;
 use PayonePayment\DataAbstractionLayer\Extension\PayonePaymentOrderTransactionExtension;
-use PayonePayment\Provider\Payone\Dto\SecuredInvoice\InvoiceDocumentDataDto;
 use PayonePayment\Provider\Payone\Extension\SecuredInvoice\SecuredInvoiceDocumentDataExtension;
 use PayonePayment\Provider\Payone\PaymentMethod\SecuredInvoicePaymentMethod;
 use Shopware\Core\Checkout\Document\Event\InvoiceOrdersEvent;
+use Shopware\Core\Checkout\Order\Aggregate\OrderTransaction\OrderTransactionEntity;
 use Shopware\Core\Checkout\Order\OrderEntity;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
@@ -28,53 +28,42 @@ readonly class InvoiceRendererEventListener implements EventSubscriberInterface
         $orders = $event->getOrders();
 
         foreach ($orders as $order) {
-            if ($this->hasSecuredInvoiceTransaction($order)) {
-                $this->addInvoiceDocumentExtension($order);
+            $transaction = $order->getTransactions()?->last();
+            if (
+                $transaction instanceof OrderTransactionEntity
+                && SecuredInvoicePaymentMethod::UUID === $transaction->getPaymentMethodId()
+            ) {
+                $this->addInvoiceDocumentExtension($order, $transaction);
             }
         }
     }
 
-    private function hasSecuredInvoiceTransaction(OrderEntity $order): bool
+    private function addInvoiceDocumentExtension(OrderEntity $order, OrderTransactionEntity $transaction): void
     {
-        if ($order->getTransactions()) {
-            foreach ($order->getTransactions() as $transaction) {
-                if (SecuredInvoicePaymentMethod::UUID === $transaction->getPaymentMethodId()) {
-                    return true;
-                }
-            }
+        $extensions = $transaction->getExtensions();
+        $extension  = $extensions[PayonePaymentOrderTransactionExtension::NAME] ?? null;
+        if (!$extension instanceof PayonePaymentOrderTransactionDataEntity) {
+            return;
         }
 
-        return false;
-    }
-
-    private function addInvoiceDocumentExtension(OrderEntity $order): void
-    {
-        $data = new SecuredInvoiceDocumentDataExtension();
-
-        foreach ($order->getTransactions() as $transaction) {
-            $extensions = $transaction->getExtensions();
-            $extension  = $extensions[PayonePaymentOrderTransactionExtension::NAME] ?? null;
-
-            if ($extension instanceof PayonePaymentOrderTransactionDataEntity) {
-                $txData = $extension->getTransactionData() ?? [];
-                foreach ($txData as $txDatum) {
-                    $clearing = $txDatum['response']['clearing'] ?? [];
-                    if ([] === $clearing) {
-                        continue;
-                    }
-
-                    $dto = new InvoiceDocumentDataDto(
-                        $clearing['BankAccount']['BankAccountHolder'] ?? null,
-                        $clearing['BankAccount']['Iban'] ?? null,
-                        $clearing['BankAccount']['Bic'] ?? null,
-                        $clearing['DueDate'] ?? null,
-                        $clearing['Reference'] ?? null,
-                    );
-                    $data->addDocumentData($dto);
-                }
-            }
+        $txData = $extension->getTransactionData() ?? [];
+        if ([] === $txData) {
+            return;
         }
 
-        $order->addExtension(SecuredInvoiceDocumentDataExtension::EXTENSION_NAME, $data);
+        $txDatum  = end($txData);
+        $clearing = $txDatum['response']['clearing'] ?? [];
+        if ([] === $clearing) {
+            return;
+        }
+
+        $ext = new SecuredInvoiceDocumentDataExtension(
+            $clearing['BankAccount']['BankAccountHolder'] ?? null,
+            $clearing['BankAccount']['Iban'] ?? null,
+            $clearing['BankAccount']['Bic'] ?? null,
+            $clearing['DueDate'] ?? null,
+            $clearing['Reference'] ?? null,
+        );
+        $order->addExtension(SecuredInvoiceDocumentDataExtension::EXTENSION_NAME, $ext);
     }
 }
