+++ src/Provider/Payone/Extension/SecuredInvoice/SecuredInvoiceDocumentDataExtension.php
@@ -4,31 +4,18 @@ declare(strict_types=1);
 
 namespace PayonePayment\Provider\Payone\Extension\SecuredInvoice;
 
-use PayonePayment\Provider\Payone\Dto\SecuredInvoice\InvoiceDocumentDataDto;
 use Shopware\Core\Framework\Struct\Struct;
 
 class SecuredInvoiceDocumentDataExtension extends Struct
 {
     final public const EXTENSION_NAME = 'payone_secured_invoice_data';
 
-    private array $documentData = [];
-
-    public function __construct()
-    {
-    }
-
-    public function addDocumentData(InvoiceDocumentDataDto $documentData): void
-    {
-        $this->documentData[] = $documentData;
-    }
-
-    public function getDocumentData(): array
-    {
-        return $this->documentData;
-    }
-
-    public function setDocumentData(array $documentData): void
-    {
-        $this->documentData = $documentData;
+    public function __construct(
+        public string|null $accountHolder,
+        public string|null $iban,
+        public string|null $bic,
+        public string|null $dueDate,
+        public string|null $reference,
+    ) {
     }
 }
