+++ src/Resources/views/documents/invoice.html.twig
@@ -132,41 +132,40 @@
             <br>
             <strong>{{ 'PayonePayment.securedInvoice.document.invoicingDescription' | trans }}</strong><br>
 
-            {% for securedInvoiceData in order.extensions[securedInvoiceDataExtension].documentData %}
-                {% if securedInvoiceData.accountHolder %}
-                    {% block document_payment_payone_secured_invoice_holder %}
-                        <br>
-                        {{ 'PayonePayment.securedInvoice.document.holder' | trans({'%holder%': securedInvoiceData.accountHolder}) }}
-                    {% endblock %}
-                {% endif %}
-
-                {% block document_payment_payone_secured_invoice_iban %}
+            {% set securedInvoiceData = order.extensions[securedInvoiceDataExtension] %}
+            {% if securedInvoiceData.accountHolder %}
+                {% block document_payment_payone_secured_invoice_holder %}
                     <br>
-                    {{ 'PayonePayment.securedInvoice.document.iban' | trans({'%iban%': securedInvoiceData.iban}) }}
+                    {{ 'PayonePayment.securedInvoice.document.holder' | trans({'%holder%': securedInvoiceData.accountHolder}) }}
                 {% endblock %}
+            {% endif %}
 
-                {% if securedInvoiceData.bic %}
-                    {% block document_payment_payone_secured_invoice_bic %}
-                        <br>
-                        {{ 'PayonePayment.securedInvoice.document.bic' | trans({'%bic%': securedInvoiceData.bic}) }}
-                    {% endblock %}
-                {% endif %}
+            {% block document_payment_payone_secured_invoice_iban %}
+                <br>
+                {{ 'PayonePayment.securedInvoice.document.iban' | trans({'%iban%': securedInvoiceData.iban}) }}
+            {% endblock %}
 
-                {% if securedInvoiceData.reference %}
-                    {% block document_payment_payone_secured_invoice_reference %}
-                        <br>
-                        {{ 'PayonePayment.securedInvoice.document.reference' | trans({'%reference%': securedInvoiceData.reference}) }}
-                    {% endblock %}
-                {% endif %}
+            {% if securedInvoiceData.bic %}
+                {% block document_payment_payone_secured_invoice_bic %}
+                    <br>
+                    {{ 'PayonePayment.securedInvoice.document.bic' | trans({'%bic%': securedInvoiceData.bic}) }}
+                {% endblock %}
+            {% endif %}
 
-                {% if securedInvoiceData.dueDate %}
-                    {% block document_payment_payone_secured_invoice_due_date %}
-                        <br>
-                        {% set formattedDueDate = securedInvoiceData.dueDate|slice(0, 4) ~ "-" ~ securedInvoiceData.dueDate|slice(4, 2) ~ "-" ~ securedInvoiceData.dueDate|slice(6, 2) %}
-                        {{ 'PayonePayment.securedInvoice.document.dueDate' | trans({'%dueDate%': formattedDueDate|format_date('medium')}) }}
-                    {% endblock %}
-                {% endif %}
-            {% endfor %}
+            {% if securedInvoiceData.reference %}
+                {% block document_payment_payone_secured_invoice_reference %}
+                    <br>
+                    {{ 'PayonePayment.securedInvoice.document.reference' | trans({'%reference%': securedInvoiceData.reference}) }}
+                {% endblock %}
+            {% endif %}
+
+            {% if securedInvoiceData.dueDate %}
+                {% block document_payment_payone_secured_invoice_due_date %}
+                    <br>
+                    {% set formattedDueDate = securedInvoiceData.dueDate|slice(0, 4) ~ "-" ~ securedInvoiceData.dueDate|slice(4, 2) ~ "-" ~ securedInvoiceData.dueDate|slice(6, 2) %}
+                    {{ 'PayonePayment.securedInvoice.document.dueDate' | trans({'%dueDate%': formattedDueDate|format_date('medium', locale=locale)}) }}
+                {% endblock %}
+            {% endif %}
 
             <br><br><strong><em>{{ 'PayonePayment.securedInvoice.document.accountNotice' | trans }}</em></strong><br>
         {% endif %}
